requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
asgiref==3.8.1
asyncer==0.0.7
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
bidict==0.23.1
build==1.2.2
cachetools==5.5.0
camelot-py==0.9.0
certifi==2024.8.30
cffi==1.17.1
chainlit==1.2.0
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.6
chromadb==0.5.7
ci-info==0.3.0
click==8.1.7
coloredlogs==15.0.1
configobj==5.0.8
configparser==7.1.0
cryptography==43.0.1
dataclasses-json==0.6.7
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
et-xmlfile==1.1.0
etelemetry==0.3.1
fastapi==0.112.4
filelock==3.16.1
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==24.3.25
frontend==0.0.3
frozenlist==1.4.1
fsspec==2024.9.0
google-auth==2.34.0
googleapis-common-protos==1.65.0
greenlet==3.1.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
inexactsearch==1.0.2
isodate==0.6.1
itsdangerous==2.2.0
jiter==0.5.0
jq==1.8.0
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==30.1.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.1
langchain-mongodb==0.2.0
langchain-openai==0.2.0
langchain-text-splitters==0.3.0
langsmith==0.1.123
Lazify==0.4.0
literalai==0.0.607
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
marshmallow==3.22.0
mdurl==0.1.2
mmh3==5.0.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.47.0
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
posthog==3.6.6
protobuf==4.25.5
prov==2.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydot==3.0.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.9.1
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
PyPDF2==3.0.1
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.4
pytils==0.4.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.9
scipy==1.14.1
shellingham==1.5.4
silpa_common==0.3
simple-websocket==1.0.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
soundex==1.1.3
spellchecker==0.4
SQLAlchemy==2.0.35
starlette==0.37.2
sympy==1.13.3
syncer==2.0.3
tenacity==8.5.0
tesseract==0.1.3
tiktoken==0.7.0
tokenizers==0.20.0
tomli==2.0.1
tools==0.1.9
tqdm==4.66.5
traits==6.3.2
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uptrace==1.26.0
urllib3==2.2.3
uvicorn==0.25.0
uvloop==0.20.0
watchfiles==0.20.0
websocket-client==1.8.0
websockets==13.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.11.1
zipp==3.20.2
